Move objectmonitorbbb to plugins directory
