+++ pom.xml
@@ -266,6 +266,12 @@
 
             <dependency>
                 <groupId>${project.groupId}</groupId>
+                <artifactId>qcc-plugin-objectmonitor</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>${project.groupId}</groupId>
                 <artifactId>qcc-plugin-optimization</artifactId>
                 <version>${project.version}</version>
             </dependency>
